Minor refactoring on the client side
